added javadocs

+++ Task_2_3_1/src/main/java/ru/nsu/kislitsyn/snake/Timer.java
@@ -2,26 +2,44 @@ package ru.nsu.kislitsyn.snake;
 
 import javafx.animation.AnimationTimer;
 
+/**
+ * Extends animation timer to add speed field for adjustment of animation velocity.
+ */
 public abstract class Timer extends AnimationTimer {
-    private long speed;
+    private double speed;
     private long prevTime;
 
-    public void setSpeed(long speed) {
+    /**
+     * Setter for the speed.
+     */
+    public void setSpeed(double speed) {
         this.speed = speed;
     }
 
-    public long getSpeed() {
+    /**
+     * Getter for the speed.
+     */
+    public double getSpeed() {
         return speed;
     }
 
+    /**
+     * Getter for the timestamp of previous step of animation.
+     */
     public long getPrevTime() {
         return prevTime;
     }
 
+    /**
+     * Setter for the timestamp of previous step of animation.
+     */
     public void setPrevTime(long prevTime) {
         this.prevTime = prevTime;
     }
 
+    /**
+     * Simple constructor.
+     */
     public Timer(long speed) {
         this.speed = speed;
     }
